1.02
Added as_abbrev, which is used when working with the megadataset. It
abbreviates the names of countries to the ISO-3 format, with custom
additions for countries or regions that lack an ISO-3 abbreviation.

Adde semi_par() which calculates semi-partial correlations, possibly
using weights. This was not possible before as far as I know.