Various
str_detect_replace. Always wanted to use mapvalues with regex? Now you
can.

MOD_summary. Bugfix for when users use chr types as factors. This gave
an error. Now it gives a warning. Also added an error if a user tries to
use a higher order variable because this is currently not supported. And
also added sample size and renamed the list output, so that may break
some code.