2016.04.27
Fixed docu for a number of functions.

FA_residuals() had a statistical error. The residuals were not
standardized before they were used. This resulted in their scales
affecting the results. Unfortunately, this affects a number of
publications using this function.

lm_beta_matrix, renamed to MOD_APSLM to be consistent with the prefixes.
Removed the VIF statistics because this was incorrect. Added R and R
adj. (multiple R).