June update.
There are new bugs in psych and VIM that cause errors in some functions. They have been reported, but unfortunately, for now the functions may not work. Affected functions are those that rely on psych::fa and VIM::irmi.

Added ggrepel as dependency.

No longer have a depdenency on XLConnect. This is annoying package that requires Java. Now I use readxl, which doesn't, and it from tidyverse.

Slight update to political units file.

Apparently, some file renaming. Due to switching to linux OS in which filenames are case-sensitive.