new function: NA_to_X() to replace NAs with a specified value.
new function: add_item_associations_with_criterion_var(), which inputs mirt item stats and a dataset, to compute the necessarily criterion relationships for use with Jensen's method with item level data.

tests for above

A fix for summarize_models() which no longer fails when names are only partially given for the list of models.